test: Move concurrency stress test to JVM-only directory
java.util.concurrent.Executors isn't available on Scala.js, so the
shared cross-compiled tests fail to link. Split the multi-threaded
regression into uni/.jvm/src/test where it can use real threads.
